+++ oneflow/user/kernels/upsample_nearest_kernel.cpp
@@ -26,7 +26,7 @@ template<typename T>
 static void UpsampleNearest1DForward(const int64_t elem_cnt, const T* in_dptr,
                                      NdIndexOffsetHelper<int64_t, 3> in_helper,
                                      NdIndexOffsetHelper<int64_t, 3> out_helper,
-                                     const int64_t in_height, const float scale_factor,
+                                     const int64_t in_height, const double scale_factor,
                                      T* out_dptr) {
   for (int64_t index = 0; index < elem_cnt; ++index) {
     int64_t n, c, h;
@@ -40,7 +40,7 @@ template<typename T>
 static void UpsampleNearest1DBackward(const int64_t elem_cnt, const T* dy_dptr,
                                       NdIndexOffsetHelper<int64_t, 3> dy_helper,
                                       NdIndexOffsetHelper<int64_t, 3> dx_helper,
-                                      const int64_t in_height, const float scale_factor,
+                                      const int64_t in_height, const double scale_factor,
                                       T* dx_dptr) {
   for (int64_t index = 0; index < elem_cnt; ++index) {
     int64_t n, c, h;
@@ -55,7 +55,7 @@ static void UpsampleNearest2DForward(const int64_t elem_cnt, const T* in_dptr,
                                      NdIndexOffsetHelper<int64_t, 4> in_helper,
                                      NdIndexOffsetHelper<int64_t, 4> out_helper,
                                      const int64_t in_height, const int64_t in_width,
-                                     const float scale_h, const float scale_w, T* out_dptr) {
+                                     const double scale_h, const double scale_w, T* out_dptr) {
   for (int64_t index = 0; index < elem_cnt; ++index) {
     int64_t n, c, h, w;
     out_helper.OffsetToNdIndex(index, n, c, h, w);
@@ -70,7 +70,7 @@ static void UpsampleNearest2DBackward(const int64_t elem_cnt, const T* dy_dptr,
                                       NdIndexOffsetHelper<int64_t, 4> dy_helper,
                                       NdIndexOffsetHelper<int64_t, 4> dx_helper,
                                       const int64_t dx_height, const int64_t dx_width,
-                                      const float scale_h, const float scale_w, T* dx_dptr) {
+                                      const double scale_h, const double scale_w, T* dx_dptr) {
   for (int64_t index = 0; index < elem_cnt; ++index) {
     int64_t n, c, h, w;
     dy_helper.OffsetToNdIndex(index, n, c, h, w);
@@ -132,7 +132,7 @@ class UpsampleNearest1DCPUKernel final : public user_op::OpKernel {
     const int64_t channels = x_tensor->shape().At(1);
     const int64_t in_height = x_tensor->shape().At(2);
     const int64_t out_height = y_tensor->shape().At(2);
-    float height_scale = ctx->Attr<float>("scale_factor");
+    double height_scale = ctx->Attr<double>("scale_factor");
     std::vector<int64_t> output_size = ctx->Attr<std::vector<int64_t>>("output_size");
     if (output_size.size()) { height_scale = out_height * 1.0 / in_height; }
 
@@ -171,7 +171,7 @@ class UpsampleNearestGrad1DCPUKernel final : public user_op::OpKernel {
     const int64_t channels = dx_tensor->shape().At(1);
     const int64_t in_height = dx_tensor->shape().At(2);
     const int64_t out_height = dy_tensor->shape().At(2);
-    float height_scale = ctx->Attr<float>("scale_factor");
+    double height_scale = ctx->Attr<double>("scale_factor");
     std::vector<int64_t> output_size = ctx->Attr<std::vector<int64_t>>("output_size");
     if (output_size.size()) { height_scale = out_height * 1.0 / in_height; }
     if (in_height == out_height) {
@@ -221,10 +221,10 @@ class UpsampleNearest2DCPUKernel final : public user_op::OpKernel {
     const int64_t out_height = y_tensor->shape().At(2);
     const int64_t out_width = y_tensor->shape().At(3);
 
-    float height_scale = ctx->Attr<float>("height_scale");
+    double height_scale = ctx->Attr<double>("height_scale");
     std::vector<int64_t> output_size = ctx->Attr<std::vector<int64_t>>("output_size");
     if (output_size.size()) { height_scale = out_height * 1.0 / in_height; }
-    float width_scale = ctx->Attr<float>("width_scale");
+    double width_scale = ctx->Attr<double>("width_scale");
     if (output_size.size()) { width_scale = out_width * 1.0 / in_width; }
 
     const int64_t elem_cnt = y_tensor->shape().elem_cnt();
@@ -266,10 +266,10 @@ class UpsampleNearest2DGradCPUKernel final : public user_op::OpKernel {
     const int64_t out_height = dy_tensor->shape().At(2);
     const int64_t out_width = dy_tensor->shape().At(3);
 
-    float height_scale = ctx->Attr<float>("height_scale");
+    double height_scale = ctx->Attr<double>("height_scale");
     std::vector<int64_t> output_size = ctx->Attr<std::vector<int64_t>>("output_size");
     if (output_size.size()) { height_scale = out_height * 1.0 / in_height; }
-    float width_scale = ctx->Attr<float>("width_scale");
+    double width_scale = ctx->Attr<double>("width_scale");
     if (output_size.size()) { width_scale = out_width * 1.0 / in_width; }
     const int64_t elem_cnt = dy_tensor->shape().elem_cnt();
